requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
anthropic==0.68.0
voyageai==0.3.4
faiss-cpu==1.12.0
numpy==2.3.2
pydantic==2.11.7
pypdf
sentence-transformers
